cloud-functions/firebase-data-api/requirements.txt
functions-framework==3.5.0
Flask==3.0.0
google-cloud-firestore==2.14.0
